# specs/001-facial-expression-editing/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Flex — Facial Expression Editing API
  version: 1.0.0
  description: >
    Prompt-guided facial expression editing for animated characters. Edits are
    confined to a selected face region; off-mask pixels are preserved byte-for-byte
    by compositing. Backend exposes only this REST contract to the frontend.
servers:
  - url: /api/v1

paths:
  /health:
    get:
      summary: Liveness/readiness (models loaded)
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status, device, model_id, models_ready]
                properties:
                  status: { type: string, example: ok }
                  device: { type: string, example: cuda }
                  model_id: { type: string }
                  models_ready: { type: boolean }

  /presets:
    get:
      summary: Supported expression presets and regions
      operationId: listPresets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [expressions, regions]
                properties:
                  expressions:
                    type: array
                    items: { type: string, enum: [smile, angry, sad, surprised] }
                  regions:
                    type: array
                    items: { $ref: "#/components/schemas/FacialRegion" }

  /images:
    post:
      summary: Upload a character image
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Image" }
        "422": { $ref: "#/components/responses/DomainError" }

  /images/{image_id}:
    get:
      summary: Image metadata
      operationId: getImage
      parameters: [ { $ref: "#/components/parameters/ImageId" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Image" }
        "404": { $ref: "#/components/responses/NotFound" }

  /images/{image_id}/file:
    get:
      summary: Raw image bytes
      operationId: getImageFile
      parameters: [ { $ref: "#/components/parameters/ImageId" } ]
      responses:
        "200":
          description: OK
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }

  /images/{image_id}/segments:
    post:
      summary: Run automatic face segmentation
      operationId: segmentImage
      parameters: [ { $ref: "#/components/parameters/ImageId" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [regions]
                properties:
                  regions:
                    type: object
                    additionalProperties: { $ref: "#/components/schemas/RegionMask" }
                    description: keyed by FacialRegion (face, eyes, mouth, eyebrows)
        "404": { $ref: "#/components/responses/NotFound" }

  /images/{image_id}/segments/{region}/file:
    get:
      summary: Region mask PNG (white=region)
      operationId: getSegmentFile
      parameters:
        - { $ref: "#/components/parameters/ImageId" }
        - name: region
          in: path
          required: true
          schema: { $ref: "#/components/schemas/FacialRegion" }
      responses:
        "200":
          description: OK
          content:
            image/png: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/NotFound" }

  /edits:
    post:
      summary: Create an edit job (async)
      operationId: createEdit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEditRequest" }
      responses:
        "202":
          description: Accepted — job queued
          content:
            application/json:
              schema:
                type: object
                required: [edit_id, status]
                properties:
                  edit_id: { type: string }
                  status: { $ref: "#/components/schemas/JobStatus" }
        "422": { $ref: "#/components/responses/DomainError" }

  /edits/{edit_id}:
    get:
      summary: Job status (+ result refs when completed)
      operationId: getEdit
      parameters: [ { $ref: "#/components/parameters/EditId" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditStatus" }
        "404": { $ref: "#/components/responses/NotFound" }

  /edits/{edit_id}/comparison:
    get:
      summary: Original + edited + evaluation metrics
      operationId: getComparison
      parameters: [ { $ref: "#/components/parameters/EditId" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comparison" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Job not yet completed

  /edits/{edit_id}/result/file:
    get:
      summary: Edited image bytes
      operationId: getResultFile
      parameters: [ { $ref: "#/components/parameters/EditId" } ]
      responses:
        "200":
          description: OK
          content:
            image/png: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/NotFound" }

components:
  parameters:
    ImageId:
      name: image_id
      in: path
      required: true
      schema: { type: string }
    EditId:
      name: edit_id
      in: path
      required: true
      schema: { type: string }

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    DomainError:
      description: Validation / domain rule violation (FR-008)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:
    FacialRegion:
      type: string
      enum: [face, eyes, mouth, eyebrows, custom]

    JobStatus:
      type: string
      enum: [processing, completed, failed]

    Image:
      type: object
      required: [image_id, width, height, url]
      properties:
        image_id: { type: string }
        width: { type: integer }
        height: { type: integer }
        url: { type: string }

    RegionMask:
      type: object
      required: [url, bbox, area]
      properties:
        url: { type: string }
        bbox:
          type: array
          items: { type: integer }
          minItems: 4
          maxItems: 4
          description: "[x, y, w, h]"
        area: { type: integer }

    EditParams:
      type: object
      properties:
        strength: { type: number, default: 0.99, exclusiveMinimum: 0, exclusiveMaximum: 1 }
        steps: { type: integer, default: 25, minimum: 10, maximum: 50 }
        guidance: { type: number, default: 8.0, minimum: 1, maximum: 20 }
        seed: { type: integer, minimum: 0, nullable: true }

    CreateEditRequest:
      type: object
      required: [image_id]
      description: >
        Exactly one selection of `box` (primary), `mask`, or `region`; and exactly one
        of `prompt`|`preset`. `box` and `mask` are mutually exclusive. Violations → 422
        (FR-008).
      properties:
        image_id: { type: string }
        box:
          type: array
          items: { type: integer }
          minItems: 4
          maxItems: 4
          description: "dragged rectangle [x, y, w, h] in image pixels (primary UX)"
        region: { $ref: "#/components/schemas/FacialRegion" }
        mask:
          type: string
          description: optional base64 PNG brush mask (white=edit)
        prompt: { type: string, minLength: 1 }
        preset: { type: string, enum: [smile, angry, sad, surprised] }
        params: { $ref: "#/components/schemas/EditParams" }

    Metrics:
      type: object
      required: [clip_similarity_in_mask, edit_success, edit_success_score, identity_preserved, latency_ms]
      properties:
        clip_similarity_in_mask: { type: number }
        edit_success: { type: boolean }
        edit_success_score: { type: number }
        identity_preserved: { type: boolean }
        latency_ms: { type: integer }

    EditStatus:
      type: object
      required: [edit_id, status]
      properties:
        edit_id: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        result:
          type: object
          nullable: true
          properties:
            result_image_id: { type: string }
            url: { type: string }
        metrics:
          oneOf:
            - { $ref: "#/components/schemas/Metrics" }
            - { type: "null" }
        error: { type: string, nullable: true }

    Comparison:
      type: object
      required: [edit_id, status, original, edited, region, prompt, metrics]
      properties:
        edit_id: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        original:
          type: object
          properties:
            image_id: { type: string }
            url: { type: string }
        edited:
          type: object
          properties:
            image_id: { type: string }
            url: { type: string }
        region: { $ref: "#/components/schemas/FacialRegion" }
        prompt: { type: string }
        metrics: { $ref: "#/components/schemas/Metrics" }

    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code: { type: string }